// src/app/models/produto.ts
export interface Produto {
   id?:                any;
   dataAbertura?:   string;
   dataFechamento?: string;
   titulo:          string;
   descricao:       string;
   preco:              any;
   estoq:           string;
   categoria:       string;
   operador:           any;
   nomeOperador:    string;
}